# backend/internal/httpapi/apidocs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API do LogN. Erro de rota que o app lê sai como {\"code\", \"message\"}, e o código é contrato.",
        "title": "LogN API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/v1/auth/request-otp": {
            "post": {
                "description": "O e-mail sai na língua de `Accept-Language`. O reenvio tem intervalo mínimo e teto por IP e global (`otp_resend_too_soon`, com `Retry-After`).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Pede um OTP por e-mail",
                "parameters": [
                    {
                        "description": "E-mail e propósito (verify_email ou reset_password)",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/httpapi.RequestOTPPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "otp_sent",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "status": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "invalid_request, invalid_email",
                        "schema": {
                            "$ref": "#/definitions/httpapi.apiError"
                        }
                    },
                    "429": {
                        "description": "otp_resend_too_soon, otp_locked, rate_limited",
                        "schema": {
                            "$ref": "#/definitions/httpapi.apiError"
                        }
                    },
                    "500": {
                        "description": "internal",
                        "schema": {
                            "$ref": "#/definitions/httpapi.apiError"
                        }
                    }
                }
            }
        },
...
        "/api/v1/auth/verify-otp": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Confere um OTP",
                "parameters": [
                    {
                        "description": "E-mail, código e propósito",
                        "name": "otp",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/httpapi.VerifyOTPPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "otp_verified",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "status": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "invalid_request",
                        "schema": {
                            "$ref": "#/definitions/httpapi.apiError"
                        }
                    },
                    "401": {
                        "description": "otp_invalid",
                        "schema": {
                            "$ref": "#/definitions/httpapi.apiError"
                        }
                    },
                    "429": {
                        "description": "rate_limited",
                        "schema": {
                            "$ref": "#/definitions/httpapi.apiError"
                        }
                    }
                }
            }
        },
...
        "/api/v1/legal/accept": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "legal"
                ],
                "summary": "Aceita os documentos pendentes",
                "parameters": [
                    {
                        "description": "Documentos, versões e hashes aceitos",
                        "name": "acceptance",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/httpapi.AcceptLegalRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "aceite gravado, sem corpo"
                    },
                    "400": {
                        "description": "invalid_request",
                        "schema": {
                            "$ref": "#/definitions/httpapi.apiError"
                        }
                    },
                    "401": {
                        "description": "unauthenticated",
                        "schema": {
                            "$ref": "#/definitions/httpapi.apiError"
                        }
                    },
                    "409": {
                        "description": "legal_version_outdated",
                        "schema": {
                            "$ref": "#/definitions/httpapi.apiError"
                        }
                    },
                    "429": {
                        "description": "rate_limited",
                        "schema": {
                            "$ref": "#/definitions/httpapi.apiError"
                        }
                    },
                    "500": {
                        "description": "internal",
                        "schema": {
                            "$ref": "#/definitions/httpapi.apiError"
                        }
                    }
                }
            }
        },
...
        "/api/v1/legal/pending": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "legal"
                ],
                "summary": "Aceites pendentes da conta",
                "parameters": [
                    {
                        "enum": [
                            "pt-BR",
                            "en",
                            "es"
                        ],
                        "type": "string",
                        "description": "Língua",
                        "name": "lang",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httpapi.pendingLegalResponse"
                        }
                    },
                    "401": {
                        "description": "unauthenticated",
                        "schema": {
                            "$ref": "#/definitions/httpapi.apiError"
                        }
                    },
                    "429": {
                        "description": "rate_limited",
                        "schema": {
                            "$ref": "#/definitions/httpapi.apiError"
                        }
                    },
                    "500": {
                        "description": "internal",
                        "schema": {
                            "$ref": "#/definitions/httpapi.apiError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "httpapi.AcceptLegalDocument": {
            "type": "object",
            "properties": {
                "from_version": {
                    "type": "integer"
                },
                "kind": {
                    "type": "string"
                },
                "locale": {
                    "type": "string"
                },
                "sha256": {
                    "type": "string"
                },
                "version": {
                    "type": "integer"
                }
            }
        },
        "httpapi.AcceptLegalRequest": {
            "type": "object",
            "properties": {
                "client": {
                    "type": "object",
                    "properties": {
                        "app": {
                            "type": "string"
                        },
                        "platform": {
                            "type": "string"
                        }
                    }
                },
                "documents": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/httpapi.AcceptLegalDocument"
                    }
                },
                "shown_changes": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "source": {
                    "description": "\"reaccept\" (a tela que bloqueia) ou \"notice\" (a faixa do aviso único).",
                    "type": "string"
                }
            }
        },
...
        "httpapi.RequestOTPPayload": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "purpose": {
                    "description": "\"verify_email\", \"reset_password\"",
                    "type": "string"
                }
            }
        },
...
        "httpapi.VerifyOTPPayload": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "purpose": {
                    "type": "string"
                }
            }
        },
        "httpapi.apiError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "httpapi.pendingLegalChange": {
            "type": "object",
            "properties": {
                "change": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "kind": {
                    "type": "string"
                },
                "section": {
                    "type": "string"
                },
                "summary": {
                    "type": "string"
                },
                "version": {
                    "type": "integer"
                }
            }
        },
        "httpapi.pendingLegalDocument": {
            "type": "object",
            "properties": {
                "accepted_effective_at": {
                    "type": "string"
                },
                "accepted_version": {
                    "description": "0 quando a conta nunca aceitou; aí a data vem vazia.",
                    "type": "integer"
                },
                "effective_at": {
                    "type": "string"
                },
                "kind": {
                    "type": "string"
                },
                "locale": {
                    "type": "string"
                },
                "material": {
                    "type": "boolean"
                },
                "sha256": {
                    "type": "string"
                },
                "version": {
                    "type": "integer"
                }
            }
        },
        "httpapi.pendingLegalResponse": {
            "type": "object",
            "properties": {
                "blocking": {
                    "description": "Alguma versão pulada é relevante: o app bloqueia até aceitar. Falso com\ndocumentos na lista é o aviso único.",
                    "type": "boolean"
                },
                "changes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/httpapi.pendingLegalChange"
                    }
                },
                "documents": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/httpapi.pendingLegalDocument"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Access token do login, no formato `Bearer \u003ctoken\u003e`.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}